session to keep who is logged in

+++ src/main/java/com/example/wishlist/Controller/WishlistController.java
@@ -4,6 +4,7 @@ import com.example.wishlist.Model.User;
 import com.example.wishlist.Model.Wish;
 import com.example.wishlist.Service.UserService;
 import com.example.wishlist.Service.WishlistService;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -22,17 +23,25 @@ public class WishlistController {
     private final WishlistService wishlistService;
     private final UserService userService;
 
+
     public WishlistController(WishlistService wishlistService, UserService userService) {
         this.wishlistService = wishlistService;
         this.userService = userService;
     }
 
+    private boolean isLoggedIn(HttpSession session) {
+        session.getAttribute("user");
+        return session.getAttribute("user") != null;
+    }
+
     @PostMapping("entrance")
     public String login(@RequestParam("email") String email,
                         @RequestParam("pw") String pw,
-                        Model model) {
+                        HttpSession session) {
         User user = userService.login(email, pw);
         if (user != null) {
+            session.setAttribute("user", user);
+            session.setMaxInactiveInterval(600);
             return "redirect:/wishes/" + user.getId() ;
         }
         return "redirect:/login";
@@ -40,19 +49,19 @@ public class WishlistController {
     }
 
     @GetMapping("login")
-    public String login(Model model) {
+    public String login(HttpSession session) {
         User user = new User();
-        model.addAttribute("user", user);
+        session.setAttribute("user", user);
         return "login";
     }
 
     @GetMapping("/wishes/{userId}")
-    public String getWishes(@PathVariable int userId, Model model) {
+    public String getWishes(@PathVariable int userId, Model model, HttpSession session) {
         User user = userService.getUser(userId);
         model.addAttribute("user", user);
         List<Wish> wishes = wishlistService.getWishesFromUser(userId);
         model.addAttribute("wishes", wishes);
-        return "wishlist";
+        return isLoggedIn(session) ? "wishlist" : "login";
     }
 
     @GetMapping("/createUser")
@@ -62,14 +71,15 @@ public class WishlistController {
         return "createUser";
     }
 
-    @PostMapping("deleteWish/{wishId}")
-    public String deleteWish(@PathVariable int wishId){
-        Wish wish = wishlistService.deleteWish(wishId);
-        return "redirect:/wishes/" + wish.getUserId();
+    @DeleteMapping("deleteWish/{wishId}")
+    public String deleteWish(@PathVariable int wishId, HttpSession session){
+        if (!isLoggedIn(session)){
+            return "login";
+        }
+        wishlistService.deleteWish(wishId);
+        User user = (User)session.getAttribute("user");
+        return "redirect:/wishes/" + user.getId();
     }
-
- // Du skal lave postmapping der tjekker om brugeren allerede eksistere,
-    // hvis den gør, så får man en fejlmeddelse og hvis ikke, så viderestilles man til login-siden
 
     @PostMapping("/register")
     public String registerUser(@ModelAttribute User user) {
@@ -77,29 +87,35 @@ public class WishlistController {
             return "redirect:/createUser";
         }
         userService.createUser(user);
-
         return "redirect:/login";
 
     }
 
-    @GetMapping("addWish/{userId}")
-    public String addWish(@PathVariable int userId, Model model){
+    @GetMapping("addWish")
+    public String addWish(Model model, HttpSession session){
         Wish wish = new Wish();
-        wish.setUserId(userId);
         model.addAttribute("wish", wish);
-        return "addWish";
+
+        return isLoggedIn(session) ? "addWish" : "login";
     }
 
 
 
     @PostMapping("saveWish")
-    public String saveWish(@ModelAttribute Wish wish){
+    public String saveWish(@ModelAttribute Wish wish, HttpSession session){
+        if (!isLoggedIn(session)){
+            return "login";
+        }
+        User user = (User)session.getAttribute("user");
         wishlistService.saveWish(wish);
-        return "redirect:/wishes/" + wish.getUserId();
+        return "redirect:/wishes/" + user.getId();
     }
 
-    @GetMapping("updateWish/{wishId}")
-    public String updateWish(@PathVariable int wishId, Model model){
+    @PutMapping("updateWish/{wishId}")
+    public String updateWish(@PathVariable int wishId, Model model, HttpSession session){
+        if (!isLoggedIn(session)){
+            return "login";
+        }
         Wish wish = wishlistService.getWishFromWishId(wishId);
         model.addAttribute("wish", wish);
         return "updateWish";
